Align the VER tag with draft08
The tag is "VER\x00", not "VER\0xff`. Note that it is necessary to
regenerate the test vectors.

This issue was not detected by our unit test `TestIETFTags()` because
the test is incorrect. We fix this test as well.

For ZZZZ, the value Tag does not match the Ascii Representation column
of Table 2. "ZZZZ" is `0x5a5a5a5a` in hex, not `0x7a7a7a7a`:
https://github.com/wbl/roughtime-draft/issues/4