Add support for draft-ietf-ntp-roughtime-10
The only change in this draft is the SRV tag used to negotiate the root
public key. Concretely, the client sends this tag to indicate which root
public key it expects to use to verify the response; when the server
receives this tag, it will use it to guide selection of the certificate
to use in the response.

This change also moves the online private key into the `Certificate`
structure - currently this is passed separately into functions that need
both the certificate and the private key. This is intended to simplify
multi-certificate setups. Accordingly, the `NewCertificate()` takes as
input the online private key rather than the public key. We also use the
types in the `ed25519` package, rather than `[]bytes`, to distinguish
public from private keys.

Finally, encapsulate the parsed request into a type, `Request`.